Refactor to split Home.tsx and VideoContent.tsx into the following components:
1. Home.tsx
2. Home-State.tsx
3. SearchSidebar.tsx
4. SearchSidebar-State.tsx
5. VideoContent.tsx
6. VideoContent-State.tsx

By using custom hooks to hold all state data, these components are easier to read and edit.
