--- src/pages/ContentPages/VideoContent/VideoContent-functions.ts
@@ -1,56 +1,0 @@
-export function isTimestampWithinRange(resTimestamp, resCreated) {
-  // Calculate the absolute difference in milliseconds
-  const difference = Math.abs(resTimestamp - resCreated);
-
-  // 2 minutes in milliseconds
-  const twoMinutesInMilliseconds = 3 * 60 * 1000;
-
-  // Check if the difference is within 2 minutes
-  return difference <= twoMinutesInMilliseconds;
-}
-
-export function extractSigValue(metadescription) {
-  // Function to extract the substring within double asterisks
-  function extractSubstring(str) {
-    const match = str.match(/\*\*(.*?)\*\*/);
-    return match ? match[1] : null;
-  }
-
-  // Function to extract the 'sig' value
-  function extractSig(str) {
-    const regex = /sig:(.*?)(;|$)/;
-    const match = str.match(regex);
-    return match ? match[1] : null;
-  }
-
-  // Extracting the relevant substring
-  const relevantSubstring = extractSubstring(metadescription);
-
-  if (relevantSubstring) {
-    // Extracting the 'sig' value
-    return extractSig(relevantSubstring);
-  } else {
-    return null;
-  }
-}
-
-export const getPaymentInfo = async (signature: string) => {
-  try {
-    const url = `/transactions/signature/${signature}`;
-    const response = await fetch(url, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
-    // Coin payment info must be added to responseData so we can display it to the user
-    const responseData = await response.json();
-    if (responseData && !responseData.error) {
-      return responseData;
-    } else {
-      throw new Error("unable to get payment");
-    }
-  } catch (error) {
-    throw new Error("unable to get payment");
-  }
-};
